bug: fix imports broken by refactoring
