+++ backend/tests/domain/test_milestones.py
@@ -1,4 +1,4 @@
-from src.domain.milestones import achieved_milestones
+from domain.milestones import achieved_milestones
 
 
 def test_milestones() -> None:
