+++ backend/tests/domain/test_streaks.py
@@ -1,4 +1,4 @@
-from src.domain.streaks import update_streak
+from domain.streaks import update_streak
 
 
 def test_streak_increment() -> None:
