+++ backend/tests/test_energy_api.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from fastapi.testclient import TestClient
 
-from src.main import app
+from main import app
 
 client = TestClient(app)
