+++ backend/tests/test_health.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 
-from src.main import app
+from main import app
 
 client = TestClient(app)
