+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-// frontend/App.tsx
+// frontend/src/App.tsx
 
 import { NavigationContainer } from '@react-navigation/native';
 import React from 'react';
