--- mypy.ini
@@ -1,7 +1,0 @@
-[mypy]
-plugins = pydantic.mypy
-ignore_missing_imports = True
-follow_imports = silent
-namespace_packages = True
-explicit_package_bases = True
-mypy_path = backend
